update audience source builder, add new component FeatureObject

+++ backend/models/enrichment_lookalike_scores.py
@@ -1,4 +1,4 @@
-from sqlalchemy import event, ForeignKey
+from sqlalchemy import event, ForeignKey, Double
 from sqlalchemy.dialects.postgresql import UUID
 from .base import Base, create_timestamps, update_timestamps
 from sqlalchemy import UUID, ForeignKey, Float
@@ -15,7 +15,7 @@ class EnrichmentLookalikeScore(Base):
     id: Mapped[UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, unique=True, server_default="gen_random_uuid()")
     lookalike_id: Mapped[UUID] = mapped_column(UUID(as_uuid=True), ForeignKey(AudienceLookalikes.id, ondelete='cascade'))
     enrichment_user_id: Mapped[UUID] = mapped_column(UUID(as_uuid=True), ForeignKey(EnrichmentUser.id, ondelete='cascade'))
-    score: Mapped[float] = mapped_column(Float)
+    score: Mapped[Double] = mapped_column(Double)
 
 event.listen(EnrichmentLookalikeScore, "before_insert", create_timestamps)
 event.listen(EnrichmentLookalikeScore, "before_update", update_timestamps)
